tests/wpt/tests/html/semantics/scripting-1/the-script-element/execution-timing/146-href.html

<!DOCTYPE html>
<html><head>
  <title>scheduler: SVG script adding src attribute </title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="testlib/testlib.js"></script>
</head>
<div id="log"></div>
<script>var t = async_test();</script>
<svg>
<script></script>
</svg>
<script>
t.step(function() {
  var s = document.querySelector("svg > script");
  s.setAttribute("href", "scripts/include-1.js");
});
onload = t.step_func(function() {
  assert_array_equals(eventOrder, ["external script #1"]);
  t.done();
});
</script>
